Move into foreground and background folders
