+++ src/Effects.js
@@ -1,52 +1,53 @@
 const constants = require('./constants');
 
-const discoBall = require('./effects/discoBall');
-const higherPower = require('./effects/higherPower');
-const rainbow = require('./effects/rainbow');
-const flowers = require('./effects/flowers');
-const colorCycle = require('./effects/colorCycle');
-const disco = require('./effects/disco');
-const ripples = require('./effects/ripples');
-const bloomingPetals = require('./effects/bloomingPetals');
-const clouds = require('./effects/clouds');
-const frostedGrid = require('./effects/frostedGrid');
-const starburst = require('./effects/starburst');
-const diamonds = require('./effects/diamonds');
-const circles = require('./effects/circles');
-const sparkles = require('./effects/sparkles');
-const text = require('./effects/text');
-const splatter = require('./effects/splatter');
-const swirl = require('./effects/swirl');
-const spiral = require('./effects/spiral');
-const lasers = require('./effects/lasers');
-const quads = require('./effects/quads');
-const kaleidoscope = require('./effects/kaleidoscope');
-const snowflakes = require('./effects/snowflakes');
-const fireworks = require('./effects/fireworks');
-const stars = require('./effects/stars');
-const galaxy = require('./effects/galaxy');
-const growingStars = require('./effects/growingStars');
-const squiggles = require('./effects/squiggles');
-const musicWave = require('./effects/musicWave');
+// background
+const discoBall = require('./effects/background/discoBall');
+const higherPower = require('./effects/background/higherPower');
+const rainbow = require('./effects/background/rainbow');
+const flowers = require('./effects/background/flowers');
+const colorCycle = require('./effects/background/colorCycle');
+const disco = require('./effects/background/disco');
+const ripples = require('./effects/background/ripples');
+const bloomingPetals = require('./effects/background/bloomingPetals');
+const clouds = require('./effects/background/clouds');
+const frostedGrid = require('./effects/background/frostedGrid');
+const starburst = require('./effects/background/starburst');
+const diamonds = require('./effects/background/diamonds');
+const circles = require('./effects/background/circles');
+const sparkles = require('./effects/background/sparkles');
+const text = require('./effects/background/text');
+const splatter = require('./effects/background/splatter');
+const swirl = require('./effects/background/swirl');
+const spiral = require('./effects/background/spiral');
+const lasers = require('./effects/background/lasers');
+const quads = require('./effects/background/quads');
+const kaleidoscope = require('./effects/background/kaleidoscope');
+const snowflakes = require('./effects/background/snowflakes');
+const fireworks = require('./effects/background/fireworks');
+const stars = require('./effects/background/stars');
+const galaxy = require('./effects/background/galaxy');
+const growingStars = require('./effects/background/growingStars');
+const squiggles = require('./effects/background/squiggles');
+const musicWave = require('./effects/background/musicWave');
 
 // foreground
-const rain = require('./effects/rain');
-const rainingTacos = require('./effects/rainingTacos');
-const pineapples = require('./effects/pineapples');
-const spotlight = require('./effects/spotlight');
-const colorLights = require('./effects/colorLights');
-const smilingPoop = require('./effects/smilingPoop');
-const heartsRed = require('./effects/heartsRed');
-const heartsColorful = require('./effects/heartsColorful');
-const floatingRainbows = require('./effects/floatingRainbows');
-const bubbles = require('./effects/bubbles');
-const explodingStars = require('./effects/explodingStars');
-const pizzas = require('./effects/pizzas');
-const smileFace = require('./effects/smileFace');
-const confetti = require('./effects/confetti');
-const musicNotes = require('./effects/musicNotes');
-const paintDrip = require('./effects/paintDrip');
-const emojis = require('./effects/emojis');
+const rain = require('./effects/foreground/rain');
+const rainingTacos = require('./effects/foreground/rainingTacos');
+const pineapples = require('./effects/foreground/pineapples');
+const spotlight = require('./effects/foreground/spotlight');
+const colorLights = require('./effects/foreground/colorLights');
+const smilingPoop = require('./effects/foreground/smilingPoop');
+const heartsRed = require('./effects/foreground/heartsRed');
+const heartsColorful = require('./effects/foreground/heartsColorful');
+const floatingRainbows = require('./effects/foreground/floatingRainbows');
+const bubbles = require('./effects/foreground/bubbles');
+const explodingStars = require('./effects/foreground/explodingStars');
+const pizzas = require('./effects/foreground/pizzas');
+const smileFace = require('./effects/foreground/smileFace');
+const confetti = require('./effects/foreground/confetti');
+const musicNotes = require('./effects/foreground/musicNotes');
+const paintDrip = require('./effects/foreground/paintDrip');
+const emojis = require('./effects/foreground/emojis');
 
 module.exports = class Effects {
   constructor(p5, alpha, extraImages, blend, currentPalette = 'default') {
