+++ src/effects/background/bloomingPetals.js
@@ -1,6 +1,6 @@
-const constants = require('../constants');
-const {hexToRgb} = require('../utils');
-const drawPetal = require('../shapes/petal');
+const constants = require('../../constants');
+const {hexToRgb} = require('../../utils');
+const drawPetal = require('../../shapes/petal');
 
 // This effect is slightly modified from Poetry background effect 'blooming'
 // https://github.com/code-dot-org/code-dot-org/blob/381e9b93f7cbd081738dfa7adbc9e7ce4e169a0c/apps/src/p5lab/poetry/commands/backgroundEffects.js#L245
