+++ src/effects/background/colorCycle.js
