+++ src/effects/background/diamonds.js
