+++ src/effects/background/galaxy.js
