+++ src/effects/background/kaleidoscope.js
