+++ src/effects/background/lasers.js
