+++ src/effects/background/rainbow.js
