+++ src/effects/background/splatter.js
