+++ src/effects/background/squiggles.js
