+++ src/effects/background/stars.js
