+++ src/effects/background/text.js
