+++ src/effects/foreground/bubbles.js
