+++ src/effects/foreground/confetti.js
