+++ src/effects/foreground/paintDrip.js
