+++ src/effects/foreground/pizzas.js
@@ -1,14 +1,14 @@
-const drawSmiley = require('../shapes/smiley');
+const drawPizza = require('../../shapes/pizza');
 
 module.exports = function (p5, randomNumber, getInPreviewMode) {
   return {
-    smiles: [],
+    pizza: [],
     init: function () {
-      if (this.smiles.length) {
+      if (this.pizza.length) {
         return;
       }
       for (let i = 0; i < 10; i++) {
-        this.smiles.push({
+        this.pizza.push({
           x: randomNumber(25, 375),
           y: randomNumber(25, 375),
           size: randomNumber(2, 6),
@@ -18,25 +18,24 @@ module.exports = function (p5, randomNumber, getInPreviewMode) {
       }
       this.image = p5.createGraphics(100, 100);
       this.image.scale(3);
-      drawSmiley(this.image.drawingContext, 0.8);
+      drawPizza(this.image.drawingContext);
     },
     draw: function (context) {
       const centroid = this.getPreviewCustomizations().getCentroid(context);
-
-      for (let i = 0; i < this.smiles.length; i++) {
+      for (let i = 0; i < this.pizza.length; i++) {
         p5.push();
-        const smiles = this.smiles[i];
-        let scale = p5.map(centroid, 5000, 8000, 0, smiles.size);
+        const pizza = this.pizza[i];
+        let scale = p5.map(centroid, 5000, 8000, 0, pizza.size);
         scale = p5.constrain(scale, 0, 5);
-        p5.translate(smiles.x, smiles.y);
-        p5.rotate(smiles.rot);
+        p5.translate(pizza.x, pizza.y);
+        p5.rotate(pizza.rot);
         p5.scale(scale / (4 * p5.pixelDensity()));
         p5.drawingContext.drawImage(this.image.elt, 0, 0);
-        smiles.life--;
-        if (smiles.life < 0) {
-          (smiles.x = randomNumber(25, 375)),
-          smiles.y - randomNumber(25, 375),
-            (smiles.life = 200);
+        pizza.life--;
+        if (pizza.life < 0) {
+          (pizza.x = randomNumber(25, 375)),
+          pizza.y - randomNumber(25, 375),
+            (pizza.life = 200);
         }
         p5.pop();
       }
