+++ src/effects/foreground/rain.js
