+++ src/effects/foreground/spotlight.js
